Add outer sequence builder to remove unused list builder
