+++ serde_arrow/src/internal/serialization_ng/array_builder.rs
@@ -1,30 +1,15 @@
 use half::f16;
 use serde::Serialize;
 
-use crate::{
-    internal::{
-        error::fail,
-        schema::{GenericDataType, GenericField, GenericTimeUnit},
-    },
-    schema::{SerdeArrowSchema, Strategy},
-    Result,
-};
+use crate::internal::error::Result;
 
 use super::{
-    bool_builder::BoolBuilder,
-    date64_builder::Date64Builder,
-    decimal_builder::DecimalBuilder,
-    dictionary_utf8_builder::DictionaryUtf8Builder,
-    float_builder::FloatBuilder,
-    int_builder::IntBuilder,
-    list_builder::ListBuilder,
-    map_builder::MapBuilder,
-    null_builder::NullBuilder,
-    struct_builder::StructBuilder,
-    union_builder::UnionBuilder,
-    unknown_variant_builder::UnknownVariantBuilder,
-    utf8_builder::Utf8Builder,
-    utils::{Mut, SimpleSerializer},
+    bool_builder::BoolBuilder, date64_builder::Date64Builder, decimal_builder::DecimalBuilder,
+    dictionary_utf8_builder::DictionaryUtf8Builder, float_builder::FloatBuilder,
+    int_builder::IntBuilder, list_builder::ListBuilder, map_builder::MapBuilder,
+    null_builder::NullBuilder, struct_builder::StructBuilder, union_builder::UnionBuilder,
+    unknown_variant_builder::UnknownVariantBuilder, utf8_builder::Utf8Builder,
+    utils::SimpleSerializer,
 };
 
 // TODO: add outer sequence builder? (not limited by i64 limits)
@@ -88,154 +73,6 @@ macro_rules! dispatch {
 }
 
 impl ArrayBuilder {
-    pub fn new(schema: &SerdeArrowSchema) -> Result<Self> {
-        let mut struct_field = GenericField::new("item", GenericDataType::Struct, false);
-        struct_field.children = schema.fields.clone();
-        let list_field =
-            GenericField::new("outer", GenericDataType::LargeList, false).with_child(struct_field);
-
-        return Ok(Self::LargeList(ListBuilder::new(
-            list_field,
-            build_struct(&schema.fields, false)?,
-            false,
-        )));
-
-        fn build_struct(fields: &[GenericField], nullable: bool) -> Result<ArrayBuilder> {
-            use ArrayBuilder as A;
-            let mut named_fields = Vec::new();
-
-            for field in fields {
-                let builder = build_builder(field)?;
-                named_fields.push((field.name.to_owned(), builder));
-            }
-
-            Ok(A::Struct(StructBuilder::new(
-                fields.to_vec(),
-                named_fields,
-                nullable,
-            )?))
-        }
-
-        fn build_builder(field: &GenericField) -> Result<ArrayBuilder> {
-            use {ArrayBuilder as A, GenericDataType as T};
-
-            let builder = match &field.data_type {
-                T::Null => {
-                    if matches!(&field.strategy, Some(Strategy::UnknownVariant)) {
-                        A::UnknownVariant(UnknownVariantBuilder)
-                    } else {
-                        A::Null(NullBuilder::new())
-                    }
-                }
-                T::Bool => A::Bool(BoolBuilder::new(field.nullable)),
-                T::I8 => A::I8(IntBuilder::new(field.nullable)),
-                T::I16 => A::I16(IntBuilder::new(field.nullable)),
-                T::I32 => A::I32(IntBuilder::new(field.nullable)),
-                T::I64 => A::I64(IntBuilder::new(field.nullable)),
-                T::U8 => A::U8(IntBuilder::new(field.nullable)),
-                T::U16 => A::U16(IntBuilder::new(field.nullable)),
-                T::U32 => A::U32(IntBuilder::new(field.nullable)),
-                T::U64 => A::U64(IntBuilder::new(field.nullable)),
-                T::F16 => A::F16(FloatBuilder::new(field.nullable)),
-                T::F32 => A::F32(FloatBuilder::new(field.nullable)),
-                T::F64 => A::F64(FloatBuilder::new(field.nullable)),
-                T::Date64 => match field.strategy.as_ref() {
-                    Some(Strategy::NaiveStrAsDate64) => {
-                        A::Date64(Date64Builder::new(field.clone(), false, field.nullable))
-                    }
-                    Some(Strategy::UtcStrAsDate64) => {
-                        A::Date64(Date64Builder::new(field.clone(), true, field.nullable))
-                    }
-                    None => A::Date64(Date64Builder::new(field.clone(), false, field.nullable)),
-                    Some(st) => fail!("Cannot builder Date64 builder with strategy {st}"),
-                },
-                T::Timestamp(unit, tz) => {
-                    if !matches!(unit, GenericTimeUnit::Millisecond) {
-                        fail!("Only timestamps with millisecond unit are supported");
-                    }
-
-                    match tz.as_deref() {
-                        None => A::Date64(Date64Builder::new(field.clone(), false, field.nullable)),
-                        Some(tz) if tz.to_uppercase() == "UTC" => {
-                            A::Date64(Date64Builder::new(field.clone(), true, field.nullable))
-                        }
-                        Some(tz) => fail!("Timezone {tz} is not supported"),
-                    }
-                }
-                T::Decimal128(precision, scale) => {
-                    A::Decimal128(DecimalBuilder::new(*precision, *scale, field.nullable))
-                }
-                T::Utf8 => A::Utf8(Utf8Builder::new(field.nullable)),
-                T::LargeUtf8 => A::LargeUtf8(Utf8Builder::new(field.nullable)),
-                T::List => {
-                    let Some(child) = field.children.first() else {
-                        fail!("cannot build a list without an element field");
-                    };
-                    A::List(ListBuilder::new(
-                        child.clone(),
-                        build_builder(child)?,
-                        field.nullable,
-                    ))
-                }
-                T::LargeList => {
-                    let Some(child) = field.children.first() else {
-                        fail!("cannot build list without an element field");
-                    };
-                    A::LargeList(ListBuilder::new(
-                        child.clone(),
-                        build_builder(child)?,
-                        field.nullable,
-                    ))
-                }
-                T::Map => {
-                    let Some(entry_field) = field.children.first() else {
-                        fail!("Cannot build a map with an entry field");
-                    };
-                    if entry_field.data_type != T::Struct && entry_field.children.len() != 2 {
-                        fail!("Invalid child field for map: {entry_field:?}")
-                    }
-                    A::Map(MapBuilder::new(
-                        entry_field.clone(),
-                        build_builder(entry_field)?,
-                        field.nullable,
-                    ))
-                }
-                T::Struct => build_struct(&field.children, field.nullable)?,
-                T::Dictionary => {
-                    let Some(indices) = field.children.first() else {
-                        fail!("Cannot build a dictionary without index field");
-                    };
-                    let Some(values) = field.children.get(1) else {
-                        fail!("Cannot build a dictionary without values field");
-                    };
-                    if !matches!(values.data_type, T::Utf8 | T::LargeUtf8) {
-                        fail!("At the moment only string dictionaries are supported");
-                    }
-                    // TODO: figure out how arrow encodes nullability and fix this
-                    let mut indices = indices.clone();
-                    indices.nullable = field.nullable;
-
-                    A::DictionaryUtf8(DictionaryUtf8Builder::new(
-                        field.clone(),
-                        build_builder(&indices)?,
-                        build_builder(values)?,
-                    ))
-                }
-                T::Union => {
-                    let mut fields = Vec::new();
-                    for field in &field.children {
-                        fields.push(build_builder(field)?);
-                    }
-
-                    A::Union(UnionBuilder::new(field.clone(), fields)?)
-                }
-            };
-            Ok(builder)
-        }
-    }
-}
-
-impl ArrayBuilder {
     pub fn name(&self) -> &'static str {
         match self {
             Self::Null(_) => "Null",
@@ -271,25 +108,6 @@ impl ArrayBuilder {
 }
 
 impl ArrayBuilder {
-    /// Try to interpret this builder as `large_list<struct>>` and extract the contained struct fields
-    pub fn take_records(&mut self) -> Result<Vec<ArrayBuilder>> {
-        let ArrayBuilder::LargeList(inner) = self else {
-            fail!("cannot take records without an outer LargeList<..>");
-        };
-        let ArrayBuilder::Struct(inner) = inner.element.as_mut() else {
-            fail!("cannot take records without an outer LargeList<Struct>");
-        };
-
-        let builder = inner.take();
-
-        let mut result = Vec::new();
-        for (_, field) in builder.named_fields {
-            result.push(field);
-        }
-
-        Ok(result)
-    }
-
     /// Take the contained array builder, while leaving structure intact
     pub fn take(&mut self) -> ArrayBuilder {
         match self {
@@ -318,12 +136,6 @@ impl ArrayBuilder {
             Self::Union(builder) => Self::Union(builder.take()),
             Self::UnknownVariant(builder) => Self::UnknownVariant(builder.take()),
         }
-    }
-}
-
-impl ArrayBuilder {
-    pub fn extend<T: Serialize + ?Sized>(&mut self, value: &T) -> Result<()> {
-        value.serialize(Mut(self))
     }
 }
