+++ serde_arrow/src/internal/serialization_ng/mod.rs
@@ -10,6 +10,7 @@ pub mod int_builder;
 pub mod list_builder;
 pub mod map_builder;
 pub mod null_builder;
+pub mod outer_sequence_builder;
 pub mod struct_builder;
 pub mod union_builder;
 pub mod unknown_variant_builder;
@@ -18,5 +19,5 @@ pub mod utils;
 
 // #[cfg(test)]
 // mod test;
-
 pub use array_builder::ArrayBuilder;
+pub use outer_sequence_builder::OuterSequenceBuilder;
